requirements.txt
streamlit
pandas
numpy
pymupdf
Pillow